fix: remove most any types

+++ src/helpers/eventBridge.ts
@@ -1,21 +1,26 @@
+/* eslint-disable max-lines */
+import { AWSError, EventBridge as AWSEventBridge, SQS } from "aws-sdk";
+import { PromiseResult } from "aws-sdk/lib/request";
 import { AWSClient, region } from "./general";
+import { removeUndefinedMessages } from "./utils/removeUndefinedMessages";
 
 export default class EventBridge {
-  QueueUrl: any;
-  eventBridgeClient: any;
-  eventBridgeName: any;
-  keep: any;
-  ruleName: any;
-  sqsClient: any;
-  targetId: any;
-  async init(eventBridgeName: any) {
+  QueueUrl: string | undefined;
+  eventBridgeClient: AWSEventBridge | undefined;
+  eventBridgeName: string | undefined;
+  keep: boolean | undefined;
+  ruleName: string | undefined;
+  sqsClient: SQS | undefined;
+  targetId: string | undefined;
+
+  async init(eventBridgeName: string): Promise<void> {
     this.eventBridgeClient = new AWSClient.EventBridge();
     this.eventBridgeName = eventBridgeName;
     this.ruleName = `test-${eventBridgeName}-rule`;
     this.targetId = "1";
 
     const keepArg = process.argv.filter((x) => x.startsWith("--keep="))[0];
-    this.keep = keepArg ? keepArg.split("=")[1] : false;
+    this.keep = keepArg ? keepArg.split("=")[1] === "true" : false;
     this.sqsClient = new AWSClient.SQS();
     if (!this.keep) {
       console.info(
@@ -31,6 +36,9 @@ export default class EventBridge {
 
     this.QueueUrl = queueResult.QueueUrl;
 
+    if (this.QueueUrl === undefined) {
+      throw new Error("QueueUrl is undefined");
+    }
     const accountId = this.QueueUrl.split("/")[3];
     const sqsArn = `arn:aws:sqs:${region}:${accountId}:${queueName}`;
     const pattern = {
@@ -83,14 +91,24 @@ export default class EventBridge {
       .promise();
   }
 
-  static async build(eventBridgeName: any) {
+  static async build(eventBridgeName: string): Promise<EventBridge> {
     const eventBridge = new EventBridge();
     await eventBridge.init(eventBridgeName);
 
     return eventBridge;
   }
 
-  async publishEvent(source: any, detailType: any, detail: any) {
+  // eslint-disable-next-line max-params
+  async publishEvent(
+    source: string | undefined,
+    detailType: string | undefined,
+    detail: string | undefined
+  ): Promise<PromiseResult<AWSEventBridge.PutEventsResponse, AWSError>> {
+    if (this.eventBridgeClient === undefined) {
+      throw new Error(
+        "EventBridgeClient is undefined. You might have forgotten to use init()"
+      );
+    }
     const result = await this.eventBridgeClient
       .putEvents({
         Entries: [
@@ -109,20 +127,30 @@ export default class EventBridge {
     return result;
   }
 
-  async getEvents() {
+  async getEvents(): Promise<SQS.ReceiveMessageResult | undefined> {
+    if (this.QueueUrl === undefined) {
+      throw new Error("QueueUrl is undefined");
+    }
     // Long poll SQS queue
     const queueParams = {
       QueueUrl: this.QueueUrl,
       WaitTimeSeconds: 5,
     };
-
+    if (this.sqsClient === undefined) {
+      throw new Error(
+        "SQSClient is undefined. You might have forgotten to use init()"
+      );
+    }
     const result = await this.sqsClient.receiveMessage(queueParams).promise();
 
-    const messageHandlers = result.Messages?.map((message: any) => ({
-      Id: message.MessageId,
-      ReceiptHandle: message.ReceiptHandle,
-    }));
-    if (messageHandlers.length > 0) {
+    const messageHandlers = removeUndefinedMessages(
+      result.Messages?.map((message: SQS.Message) => ({
+        Id: message.MessageId,
+        ReceiptHandle: message.ReceiptHandle,
+      }))
+    );
+
+    if (messageHandlers !== undefined && messageHandlers.length > 0) {
       await this.sqsClient
         .deleteMessageBatch({
           Entries: messageHandlers,
@@ -134,7 +162,15 @@ export default class EventBridge {
     return result;
   }
 
-  async clear() {
+  async clear(): Promise<any> {
+    if (this.sqsClient === undefined) {
+      throw new Error(
+        "SQSClient is undefined. You might have forgotten to use init()"
+      );
+    }
+    if (this.QueueUrl === undefined) {
+      throw new Error("QueueUrl is undefined");
+    }
     const result = await this.sqsClient
       .purgeQueue({
         QueueUrl: this.QueueUrl,
@@ -144,14 +180,44 @@ export default class EventBridge {
     return result;
   }
 
-  async destroy() {
+  async destroy(): Promise<boolean> {
+    if (this.keep === undefined) {
+      throw new Error(
+        "keep is undefined. You might have forgotten to use init()"
+      );
+    }
     if (!this.keep) {
+      if (this.sqsClient === undefined) {
+        throw new Error(
+          "SQSClient is undefined. You might have forgotten to use init()"
+        );
+      }
+      if (this.QueueUrl === undefined) {
+        throw new Error("QueueUrl is undefined");
+      }
+
       await this.sqsClient
         .deleteQueue({
           QueueUrl: this.QueueUrl,
         })
         .promise();
 
+      if (this.eventBridgeClient === undefined) {
+        throw new Error(
+          "EventBridgeClient is undefined. You might have forgotten to use init()"
+        );
+      }
+
+      if (this.targetId === undefined) {
+        throw new Error(
+          "targetId is undefined. You might have forgotten to use init()"
+        );
+      }
+      if (this.ruleName === undefined) {
+        throw new Error(
+          "ruleName is undefined. You might have forgotten to use init()"
+        );
+      }
       await this.eventBridgeClient
         .removeTargets({
           Ids: [this.targetId],
